Conflicts:
	README-merlin.txt

+++ release/src/router/www/Advanced_WAdvanced_Content.asp
@@ -167,16 +167,22 @@ function initial(){
 
 	$("wl_rate").style.display = "none";
 
-	if(!Rawifi_support){ // BRCM == without rawifi
+	if(Rawifi_support){
+		inputCtrl(document.form.wl_noisemitigation, 0);
+	}
+	else if(Qcawifi_support){
+		// FIXME
+		$("DLSCapable").style.display = "none";	
+		$("PktAggregate").style.display = "none";
+		inputCtrl(document.form.wl_noisemitigation, 0);
+	}else{
+		// BRCM
 		$("DLSCapable").style.display = "none";	
 		$("PktAggregate").style.display = "none";
 		
 		if('<% nvram_get("wl_unit"); %>' == '1' || sdk_6){	// MODELDEP: for Broadcom SDK 6.x model
 			inputCtrl(document.form.wl_noisemitigation, 0);
 		}
-	}
-	else{
-		inputCtrl(document.form.wl_noisemitigation, 0);
 	}
 
 	if(wifi_hw_sw_support){ //For N55U
@@ -185,22 +191,34 @@ function initial(){
 		}
 	}
 	
-	if((sdk_6 || sdk_7) && !Rawifi_support){		// for BRCM new SDK 6.x && SDK 7.x
-		inputCtrl(document.form.wl_ampdu_mpdu, 1);
-	}else{
+	// MODELDEP: for AC ser
+	if(Rawifi_support){
 		inputCtrl(document.form.wl_ampdu_mpdu, 0);
-	}
-
-	if(sdk_6 && !Rawifi_support){		// for BRCM new SDK 6.x
-		inputCtrl(document.form.wl_ack_ratio, 1);
-	}else{
 		inputCtrl(document.form.wl_ack_ratio, 0);
+	}else if(Qcawifi_support){
+		// FIXME
+		inputCtrl(document.form.wl_ampdu_mpdu, 0);
+		inputCtrl(document.form.wl_ack_ratio, 0);
+	}else{
+		if (sdk_6){
+			// for BRCM new SDK 6.x
+			inputCtrl(document.form.wl_ampdu_mpdu, 1);
+			inputCtrl(document.form.wl_ack_ratio, 1);
+		}else if(sdk_7){
+			// for BRCM new SDK 7.x
+			inputCtrl(document.form.wl_ampdu_mpdu, 1);
+			inputCtrl(document.form.wl_ack_ratio, 0);
+		}else{
+			inputCtrl(document.form.wl_ampdu_mpdu, 0);
+			inputCtrl(document.form.wl_ack_ratio, 0);
+		}
 	}
 	
 	inputCtrl(document.form.wl_turbo_qam, 0);
 	inputCtrl(document.form.wl_txbf, 0);
 	inputCtrl(document.form.wl_itxbf, 0);
 	inputCtrl(document.form.usb_usb3, 0);
+	inputCtrl(document.form.traffic_5g, 0);
 
 	if('<% nvram_get("wl_unit"); %>' == '1' || '<% nvram_get("wl_unit"); %>' == '2'){ // 5GHz up
 		if(	based_modelid == "RT-AC3200" ||
@@ -220,7 +238,10 @@ function initial(){
 			based_modelid == "RT-AC87U" || based_modelid == "EA-AC87")
 		{
 			inputCtrl(document.form.wl_itxbf, 1);
-		}		
+		}
+				
+		if( based_modelid == "RT-AC55U")
+			inputCtrl(document.form.traffic_5g, 1);
 	}
 	else{ // 2.4GHz
 		if(	based_modelid == "RT-AC3200" ||
@@ -228,6 +249,7 @@ function initial(){
 			based_modelid == "RT-N65U" ||
 			based_modelid == "RT-AC69U" || based_modelid == "TM-AC1900" ||
 			based_modelid == "RT-AC87U" ||
+			based_modelid == "RT-AC55U" || based_modelid == "4G-AC55U" ||
 			based_modelid == "RT-AC56S" || based_modelid == "RT-AC56U" || 
 			based_modelid == "RT-AC68U" || based_modelid == "RT-AC68U_V2" || based_modelid == "DSL-AC68U")
 		{
@@ -261,7 +283,7 @@ function initial(){
 	for (var i = 0; i < mcast_rates.length; i++) {
 		if (mcast_unit == '1' && mcast_rates[i][2]) // 5Ghz && CCK
 			continue;
-		if (!Rawifi_support && mcast_rates[i][3]) // BCM && HTMIX
+		if (!Rawifi_support && !Qcawifi_support && mcast_rates[i][3]) // BCM && HTMIX
 			continue;
 		if (Rawifi_support && HtTxStream < mcast_rates[i][4]) // ralink && HtTxStream
 			continue;
@@ -385,7 +407,7 @@ function changeRSSI(_switch){
 
 function applyRule(){
 	if(validForm()){
-		if(wifi_hw_sw_support) { //For N55U
+		if(wifi_hw_sw_support && !Qcawifi_support) { //For N55U
 			document.form.wl_HW_switch.value = "0";
 			document.form.wl_HW_switch.disabled = false;
 		}
@@ -396,6 +418,11 @@ function applyRule(){
 
 		if(document.form.usb_usb3.disabled == false && document.form.usb_usb3.value != '<% nvram_get("usb_usb3"); %>'){
 			FormActions("start_apply.htm", "apply", "reboot", "<% get_default_reboot_time(); %>");
+		}
+
+		if("<% nvram_get("wl_unit"); %>" == "1" && "<% nvram_get("wl1_country_code"); %>" == "EU" && based_modelid == "RT-AC87U"){	//for EU RT-AC87U 5G Advanced setting
+			if(document.form.wl1_80211h[0].selected && "<% nvram_get("wl1_chanspec"); %>" == "0")	//Interlocking set acs_dfs="0" while disabled 802.11h and wl1_chanspec="0"(Auto)
+				document.form.acs_dfs.value = "0";
 		}
 
 		showLoading();
@@ -422,7 +449,7 @@ function validForm(){
 		return false;
 	}
 	
-	if(power_support && !Rawifi_support){
+	if(power_support && !Rawifi_support && !Qcawifi_support){
 		// MODELDEP
 		if(hw_ver.search("RTN12HP") != -1){
 		  FormActions("start_apply.htm", "apply", "set_wltxpower;reboot", "<% get_default_reboot_time(); %>");
@@ -593,7 +620,7 @@ function setFlag_TimeFiled(){
 function enable_wme_check(obj){
 	if(obj.value == "off"){
 		inputCtrl(document.form.wl_wme_no_ack, 0);
-		if(!Rawifi_support)
+		if(!Rawifi_support && !Qcawifi_support)
 			inputCtrl(document.form.wl_igs, 0);
 		
 		inputCtrl(document.form.wl_wme_apsd, 0);
@@ -605,7 +632,7 @@ function enable_wme_check(obj){
 		}else		
 			inputCtrl(document.form.wl_wme_no_ack, 1);
 		
-		if(!Rawifi_support)
+		if(!Rawifi_support && !Qcawifi_support)
 			inputCtrl(document.form.wl_igs, 1);
 
 		inputCtrl(document.form.wl_wme_apsd, 1);
@@ -700,6 +727,8 @@ function set_power(power_value){
 <input type="hidden" name="wl_HW_switch" value="<% nvram_get("wl_HW_switch"); %>" disabled>
 <input type="hidden" name="wl_TxPower" value="<% nvram_get("wl_TxPower"); %>" >
 <input type="hidden" name="wl1_80211h_orig" value="<% nvram_get("wl1_80211h"); %>" >
+<input type="hidden" name="acs_dfs" value="<% nvram_get("acs_dfs"); %>">
+<input type="hidden" name="w_Setting" value="1">
 
 <table class="content" align="center" cellpadding="0" cellspacing="0">
 	<tr>
@@ -872,7 +901,9 @@ function set_power(power_value){
 						<select name="wl_plcphdr" class="input_option">
 							<option value="long" <% nvram_match("wl_plcphdr", "long", "selected"); %>>Long</option>
 							<option value="short" <% nvram_match("wl_plcphdr", "short", "selected"); %>>Short</option>
+<!-- auto mode applicable for STA only
 							<option value="auto" <% nvram_match("wl_plcphdr", "auto", "selected"); %>><#Auto#></option>
+-->
 						</select>
 						</td>
 					</tr>
@@ -989,6 +1020,17 @@ function set_power(power_value){
 						</td>
 					</tr>
 					
+					<tr> <!-- MODELDEP: RT-AC55U -->
+						<th><a class="hintstyle" href="javascript:void(0);"">Enable accurate traffic counter</a></th>
+						<td>
+							<select name="traffic_5g" class="input_option">
+								<option value="1" <% nvram_match("traffic_5g", "1","selected"); %>><#WLANConfig11b_WirelessCtrl_button1name#></option>
+								<option value="0" <% nvram_match("traffic_5g", "0","selected"); %>><#WLANConfig11b_WirelessCtrl_buttonname#></option>
+							</select>
+						</td>
+					</tr>
+					
+
 					<!-- [MODELDEP] for Broadcom SDK 6.x -->
 					<tr id="wl_ampdu_mpdu_field">
 						<th><a class="hintstyle" href="javascript:void(0);" onClick="openHint(3,26);"><#WLANConfig11b_x_AMPDU#></a></th>
@@ -1023,8 +1065,8 @@ function set_power(power_value){
 						<th><a class="hintstyle" href="javascript:void(0);" onClick="">IEEE 802.11h support</a></th>
 						<td>
 							<select name="wl1_80211h" class="input_option">
-									<option value="0" <% nvram_match("wl1_80211h", "0","selected"); %> ><#WLANConfig11b_WirelessCtrl_buttonname#></option>
-									<option value="1" <% nvram_match("wl1_80211h", "1","selected"); %> ><#WLANConfig11b_WirelessCtrl_button1name#></option>
+									<option value="0" <% nvram_match("wl1_80211h", "0","selected"); %>><#WLANConfig11b_WirelessCtrl_buttonname#></option>
+									<option value="1" <% nvram_match("wl1_80211h", "1","selected"); %>><#WLANConfig11b_WirelessCtrl_button1name#></option>
 							</select>
 						</td>
 					</tr>
